+++ lib/sleeping_king_studios/tools/string_tools.rb
@@ -23,5 +23,30 @@ module SleepingKingStudios::Tools
     def pluralize count, single, plural
       1 == count ? single : plural
     end # method pluralize
+
+    # Converts a mixed-case string expression to a lowercase, underscore
+    # separated string.
+    #
+    # @param str [String] The string to convert.
+    #
+    # @return [String] The converted string.
+    #
+    # @see ActiveSupport::Inflector#underscore.
+    def underscore str
+      require_string! str
+
+      str = str.dup
+      str.gsub!(/([A-Z\d]+)([A-Z][a-z])/, '\1_\2')
+      str.gsub!(/([a-z\d])([A-Z])/, '\1_\2')
+      str.tr!("-", "_")
+      str.downcase!
+      str
+    end # method underscore
+
+    private
+
+    def require_string! value
+      raise ArgumentError.new('argument must be a string') unless value.is_a?(String)
+    end # method require_array
   end # module
 end # module